fix createMany
