+++ packages/twenty-server/src/engine/twenty-orm/repository/workspace.repository.ts
@@ -40,7 +40,7 @@ export class WorkspaceRepository<
   private readonly internalContext: WorkspaceInternalContext;
   private shouldBypassPermissionChecks: boolean;
   private featureFlagMap: FeatureFlagMap;
-  private objectRecordsPermissions?: ObjectRecordsPermissions;
+  public readonly objectRecordsPermissions?: ObjectRecordsPermissions;
   private authContext?: AuthContext;
   declare manager: WorkspaceEntityManager;
